Integrando o Front com o Back, Fazendo o Registro de Historico e Historico do saldo no Front

+++ src/api/Project.Domain/Project.cs
@@ -10,5 +10,5 @@ public class Project
     public int Id { get; set; }
     public string Code { get; set; }
     public string Name { get; set; }
-    public int Budget { get; set; }
+    public decimal Budget { get; set; }
 }
